requirements.txt
defusedxml==0.4.1
Django==1.8.3
django-admin-bootstrapped==2.5.5
django-allauth==0.23.0
django-bootstrap3==6.2.2
django-braces==1.8.1
django-cors-headers==1.1.0
django-debug-toolbar==1.3.2
django-embed-video==1.0.0
django-filter==0.10.0
django-oauth-toolkit==0.8.2
django-paypal==0.2.5
django-random-queryset==0.0.3
django-rest-auth==0.5.0
django-subdomains==2.0.4
django-taggit==0.17.0
django-taggit-serializer==0.1.1
djangorestframework==3.1.3
Markdown==2.6.2
mysqlclient==1.3.6
oauthlib==0.7.2
Pillow==2.9.0
psycopg2==2.5.4
python3-openid==3.0.6
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
South==1.0.2
sqlparse==0.1.16
wheel==0.24.0
